small fixes
r2695

+++ FreeBASIC/tests/compound/scope.bas
@@ -4,19 +4,19 @@ const TEST_A = -1
 const TEST_B = -2
 const TEST_C = -3
 const TEST_D = -4
-	
+
 declare sub foo	
 
-	dim shared c = TEST_C
-	dim shared d = TEST_D
+	dim shared c as integer = TEST_C
+	dim shared d as integer = TEST_D
 	
-	dim i
+	dim i as integer
 	for i = 1 to 2
 	
 		scope
-			dim a = TEST_A
+			dim a as integer = TEST_A
 			scope 
-				dim b = TEST_B
+				dim b as integer = TEST_B
 				assert( a = TEST_A )
 				scope
 					assert( b = TEST_B )
@@ -26,12 +26,12 @@ declare sub foo
 		end scope
 		
 		scope
-			dim a = 1234
+			dim a as integer = 1234
 			assert( a = 1234 )
 			scope
 				assert( c = TEST_C )
 				scope
-					dim d = 5678
+					dim d as integer = 5678
 					assert( d = 5678 )
 				end scope
 			end scope
@@ -47,13 +47,13 @@ declare sub foo
 ''::::
 sub foo
 
-	dim i
+	dim i as integer
 	for i = 1 to 2
 
 		scope
-			dim a = TEST_A
+			dim a as integer = TEST_A
 			scope 
-				dim b = TEST_B
+				dim b as integer = TEST_B
 				assert( a = TEST_A )
 				scope
 					assert( b = TEST_B )
@@ -62,12 +62,12 @@ sub foo
 		end scope
 		
 		scope
-			dim a = 1234
+			dim a as integer = 1234
 			assert( a = 1234 )
 			scope
 				assert( c = TEST_C )
 				scope
-					dim d = 5678
+					dim d as integer = 5678
 					assert( d = 5678 )
 				end scope
 			end scope
